split from smartos-live/src/img at 0308c8680915823a54bf911a69385020143ee8cb
The aim of this is to have a minimally viable imgadm for Linux hosts.  The
changeset following 0308c86 in smartos-live was the one that introduced vminfod.
That change introduces a tighter coupling between the various node components
that are delivered as part of the PI and in a way that is very specific to
SmartOS.  vminfod will be reconsidered once we are ready to take training wheels
off.